# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Pulse
  version: '1.0.0'
  description: |
    Multi-tenant notification delivery — email, push, SMS, in-app inbox and
    outbound webhooks behind one API.

    ## Authentication
    Every tenant-facing endpoint takes an API key as `Authorization: Bearer pk_live_…`
    (or the `X-API-Key` header). Only the SHA-256 of a key is stored, so a key
    is shown exactly once, at creation, and cannot be recovered afterwards.

    The `/admin/v1` surface is guarded separately by `ADMIN_TOKEN` — it creates
    the tenants that API keys belong to, so it cannot be authenticated by one.

    ## Errors
    Every error returns `{ "error": { "code", "message", "details"? } }`.
    Switch on `code`; `message` may be reworded at any time.

    ## Idempotency
    `POST /v1/notifications` honours an `Idempotency-Key` header. A repeat with
    the same key and body replays the original response and sends nothing; the
    replay carries `Idempotent-Replay: true`. The same key with a *different*
    body is a `409 IDEMPOTENCY_KEY_REUSED`.
  license:
    name: MIT
servers:
  - url: http://localhost:3100
    description: Local
  - url: https://{apiId}.execute-api.ap-south-1.amazonaws.com
    description: AWS
    variables:
      apiId:
        default: CHANGEME
security:
  - ApiKeyAuth: []
...
paths:
...
  /v1/subscribers:
    post:
      tags: [Subscribers]
      summary: Create or update a subscriber
      description: |
        Upsert, matched by `subscriberId` then `externalId`. Preferences are
        never reset by an upsert — they belong to the subscriber and change only
        through `PUT /v1/subscribers/{id}/preferences`.

        Bangladeshi phone formats (`01712345678`, `8801712345678`) are
        normalised to E.164.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpsertSubscriber' }
      responses:
        '201':
          description: The subscriber
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscriber' }
  /v1/subscribers/{subscriberId}:
    parameters:
      - $ref: '#/components/parameters/SubscriberId'
    get:
      tags: [Subscribers]
      summary: Fetch a subscriber
      responses:
        '200':
          description: The subscriber
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscriber' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Subscribers]
      summary: Update a subscriber
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpsertSubscriber' }
      responses:
        '200':
          description: The subscriber
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscriber' }
  /v1/subscribers/{subscriberId}/preferences:
    put:
      tags: [Subscribers]
      summary: Set channel and category opt-outs
      description: |
        Both maps are opt-out: a missing entry means opted in. Merged into the
        existing preferences rather than replacing them.
      parameters:
        - $ref: '#/components/parameters/SubscriberId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Preferences' }
      responses:
        '200':
          description: The resulting preferences
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Preferences' }
  /v1/subscribers/{subscriberId}/unsubscribe:
    post:
      tags: [Subscribers]
      summary: One-click unsubscribe from a channel
      parameters:
        - $ref: '#/components/parameters/SubscriberId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel]
              properties:
                channel: { $ref: '#/components/schemas/Channel' }
                resubscribe: { type: boolean, default: false }
      responses:
        '201': { description: The resulting preferences }
  /v1/subscribers/{subscriberId}/devices:
    parameters:
      - $ref: '#/components/parameters/SubscriberId'
    get:
      tags: [Subscribers]
      summary: List registered device tokens
      responses:
        '200': { description: Device tokens }
    post:
      tags: [Subscribers]
      summary: Register a device token
      description: |
        Re-registering the same token refreshes `lastSeenAt` and keeps the
        original `createdAt`. Tokens FCM reports as permanently unregistered are
        pruned automatically after a send.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, platform]
              properties:
                token: { type: string }
                platform: { type: string, enum: [ios, android, web] }
                appVersion: { type: string }
      responses:
        '201': { description: The registered device }
  /v1/subscribers/{subscriberId}/devices/{token}:
    delete:
      tags: [Subscribers]
      summary: Remove a device token
      parameters:
        - $ref: '#/components/parameters/SubscriberId'
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Removed }
  /v1/templates:
    get:
      tags: [Templates]
      summary: List the newest version of every template
      responses:
        '200': { description: Templates }
    post:
      tags: [Templates]
      summary: Publish a template version
      description: |
        Templates are append-only. Publishing an edit creates the next version
        and leaves earlier ones readable, so a message already accepted for
        delivery can never have its text change underneath it.

        `locales.en` is required and is the per-channel fallback for every other
        locale. Bodies are Liquid — sandboxed, no arbitrary code execution.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishTemplate' }
      responses:
        '201':
          description: The published version
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Template' }
  /v1/templates/{key}:
    parameters:
      - name: key
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Templates]
      summary: Fetch a template
      parameters:
        - name: version
          in: query
          description: Omit for the newest version.
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: The template
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Template' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Templates]
      summary: Publish the next version of this key
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PublishTemplate' }
      responses:
        '200': { description: The published version }
...
  /v1/webhooks:
    get:
      tags: [Webhooks]
      summary: List endpoints (secrets redacted)
      responses:
        '200': { description: Endpoints }
    post:
      tags: [Webhooks]
      summary: Register an outbound endpoint
      description: |
        The signing secret is returned in full exactly once, here.

        Deliveries carry `Pulse-Signature: t=<unix>,v1=<hex hmac>`, an HMAC-SHA256
        over `"<t>.<raw body>"`. Verify with a constant-time compare and reject
        timestamps older than ~5 minutes — the timestamp inside the signed
        string is what makes replay protection possible.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string, format: uri }
                events:
                  type: array
                  items: { type: string }
                  default: ['*']
      responses:
        '201': { description: The endpoint, including its secret }
  /v1/webhooks/{endpointId}:
    delete:
      tags: [Webhooks]
      summary: Remove an endpoint
      parameters:
        - name: endpointId
          in: path
          required: true
          schema: { type: string }
      responses:
        '204': { description: Removed }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer pk_live_…` (or the `X-API-Key` header)'
    AdminAuth:
      type: http
      scheme: bearer
      description: The platform `ADMIN_TOKEN`.
  parameters:
...
    SubscriberId:
      name: subscriberId
      in: path
      required: true
      schema: { type: string }
  responses:
...
    NotFound:
      description: No such resource for this tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Channel:
      type: string
      enum: [email, push, sms, inapp, webhook]
    Locale:
      type: string
      enum: [en, bn]
...
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code. Switch on this.
              examples:
                - VALIDATION_FAILED
                - INVALID_API_KEY
                - QUOTA_EXCEEDED
                - TEMPLATE_NOT_FOUND
                - IDEMPOTENCY_KEY_REUSED
                - SCHEDULE_IN_PAST
            message: { type: string }
            details: {}
...
    TemplateBodies:
      type: object
      description: Per-channel bodies. An absent channel is `skipped`, not failed.
      properties:
        email:
          type: object
          required: [subject, html]
          properties:
            subject: { type: string }
            html: { type: string }
            text:
              type: string
              description: Derived from the HTML when omitted.
        push:
          type: object
          required: [title, body]
          properties:
            title: { type: string }
            body: { type: string }
            imageUrl: { type: string }
            data:
              type: object
              additionalProperties: { type: string }
        sms:
          type: object
          required: [text]
          properties:
            text: { type: string }
        inapp:
          type: object
          required: [title, body]
          properties:
            title: { type: string }
            body: { type: string }
            deeplink: { type: string }
        webhook:
          type: object
          required: [event]
          properties:
            event: { type: string }
            payload:
              type: string
              description: Liquid rendering to JSON. Defaults to the full `data` bag.
...
    Preferences:
      type: object
      properties:
        channels:
          type: object
          additionalProperties: { type: boolean }
          description: 'Missing entry = opted in. Example: `{ "email": false }`'
        categories:
          type: object
          additionalProperties: { type: boolean }
    UpsertSubscriber:
      type: object
      properties:
        subscriberId: { type: string }
        externalId: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        locale: { $ref: '#/components/schemas/Locale' }
        timezone: { type: string, default: Asia/Dhaka }
        attributes: { type: object, additionalProperties: true }
        topics:
          type: array
          items: { type: string }
    Subscriber:
      allOf:
        - $ref: '#/components/schemas/UpsertSubscriber'
        - type: object
          properties:
            tenantId: { type: string }
            preferences: { $ref: '#/components/schemas/Preferences' }
            createdAt: { type: string, format: date-time }
            updatedAt: { type: string, format: date-time }
    PublishTemplate:
      type: object
      required: [key, name, locales]
      properties:
        key: { type: string }
        name: { type: string }
        category: { type: string, default: transactional }
        locales:
          type: object
          required: [en]
          additionalProperties: { $ref: '#/components/schemas/TemplateBodies' }
    Template:
      allOf:
        - $ref: '#/components/schemas/PublishTemplate'
        - type: object
          properties:
            tenantId: { type: string }
            version: { type: integer }
            createdAt: { type: string, format: date-time }
...